refactor: extract drain_event_stream into agent_chat/drain.rs
Split stateful async orchestration (spinner lifecycle, thinking
accumulator, Rich/Raw mode selection) from the stateless per-event
formatting in renderer::render_event.

- New drain.rs module with drain_event_stream()
- renderer.rs retains only render_event() and its tests
- Updated imports in repl.rs and agent_question.rs
