+++ crates/gglib-cli/src/handlers/agent_chat/mod.rs
@@ -4,12 +4,14 @@
 //! independently readable:
 //! - [`config`]   — resolves LLM port + MCP tools, composes an [`gglib_core::ports::AgentLoopPort`]
 //! - [`renderer`] — maps [`gglib_core::AgentEvent`] variants to terminal output
+//! - [`drain`]    — async event-stream consumer (spinner, thinking accumulator)
 //! - [`repl`]     — async REPL loop with `rustyline` + `spawn_blocking` input
 //! - [`tool_format`] — tool-result summary formatters
 //! - [`markdown`] — Markdown normalisation + termimad rendering
 //! - [`thinking_dispatch`] — `RenderContext`, thinking-event dispatch, spinner coordination
 
 pub mod config;
+pub mod drain;
 mod markdown;
 pub mod persistence;
 pub mod renderer;
